Implementación de permisos a botones

+++ resources/views/livewire/admin/notifications/index.blade.php
@@ -8,13 +8,14 @@
     <div class="max-w-7xl mx-auto pt-5 pb-10">
         <div class="space-y-2">
             <!-- Botón de nuevo -->
-            <div class="mb-6">
-                <x-jet-secondary-button wire:click="create()" title="Crear nueva notificación" class="border-[#1b396a] text-[#1b396a] hover:text-sky-500 active:text-[#1b396a] active:bg-sky-50">
-                    <x-icon.plus solid alt="sm" class="inline-block h-5 w-5" />
-                    Nueva Notificación
-                </x-jet-secondary-button>
-            </div>
-
+            @can('sendnotify.create')
+                <div class="mb-6">
+                    <x-jet-secondary-button wire:click="create()" title="Crear nueva notificación" class="border-[#1b396a] text-[#1b396a] hover:text-sky-500 active:text-[#1b396a] active:bg-sky-50">
+                        <x-icon.plus solid alt="sm" class="inline-block h-5 w-5" />
+                        Nueva Notificación
+                    </x-jet-secondary-button>
+                </div>
+            @endcan
             <!-- Opciones de tabla -->
             <div class="md:flex md:justify-between space-y-2 md:space-y-0">
                 <!-- Parte izquierda -->
@@ -28,14 +29,16 @@
                     </div>
 
                 </div>
-                <div class="mb-6">
+                @can('sendnotify.delete')
                     <div class="mb-6">
-                        <x-jet-secondary-button wire:click="delete_noti()" title="Eliminar notificaciones enviadas" class="border-red-300 text-red-700 hover:text-red-500 active:text-red-800 active:bg-red-50">
-                                <x-icon.trash d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" class="inline-block h-6 w-6"/>
-                                Eliminar Notificaciones
-                        </x-jet-secondary-button>
+                        <div class="mb-6">
+                            <x-jet-secondary-button wire:click="delete_noti()" title="Eliminar notificaciones enviadas" class="border-red-300 text-red-700 hover:text-red-500 active:text-red-800 active:bg-red-50">
+                                    <x-icon.trash d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" class="inline-block h-6 w-6"/>
+                                    Eliminar Notificaciones
+                            </x-jet-secondary-button>
+                        </div>
                     </div>
-                </div>
+                @endcan
                  <!-- Parte derecha -->
                 <div  class="mb-6">
                  <!-- Selección de paginación -->
@@ -68,12 +71,16 @@
                                 <x-table.cell class="text-center">{{ $r->role}}</x-table.cell>
                                 <x-table.cell class="text-center">{{ $r->created_at->diffForHumans()}}</x-table.cell>
                                 <x-table.cell width='200' class="whitespace-nowrap">
-                                    <button  wire:click="view({{$r->id}})" type="button" title="Ver notificación enviada" class="mr-1 px-4  bg-white hover:text-white hover:bg-amber-500 text-black font-bold border border-amber-400 rounded shadow" >
-                                        Ver
-                                    </button>
-                                    <button wire:click="delete_post('{{$r->id}}' , '{{$r->title}}')" type="button" title="Eliminar notificación" class="ml-1 px-4 bg-white hover:text-white hover:bg-red-600 text-black font-bold border border-red-400 rounded shadow">
-                                        Eliminar
-                                    </button>
+                                    @can('sendnotify.show')
+                                        <button  wire:click="view({{$r->id}})" type="button" title="Ver notificación enviada" class="mr-1 px-4  bg-white hover:text-white hover:bg-amber-500 text-black font-bold border border-amber-400 rounded shadow" >
+                                            Ver
+                                        </button>
+                                    @endcan
+                                    @can('sendnotify.delete')
+                                        <button wire:click="delete_post('{{$r->id}}' , '{{$r->title}}')" type="button" title="Eliminar notificación" class="ml-1 px-4 bg-white hover:text-white hover:bg-red-600 text-black font-bold border border-red-400 rounded shadow">
+                                            Eliminar
+                                        </button>
+                                    @endcan
                                 </x-table.cell>
                             </tr>
                     @empty
